pauta-automation/requirements.txt
# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# GUI
pywebview>=4.0

# Image processing
Pillow>=10.0.0

# HTTP client
httpx>=0.25.0

# HTML parsing
beautifulsoup4>=4.12.0

# Video download
yt-dlp>=2024.1.0

# OpenAI
openai>=1.0.0

# Packaging
pyinstaller>=6.0.0

# Testing
pytest>=7.4.0
